Initialise rtc6705 CS to high.
Currently the pin initialisation sets it low, this allows any
initialisation glitches on the DATA and CLK lines to possibly be
detected.

This PR changes the order and forces the CS high before any changes to
the DATA and CLK lines.

I know there is information out there that the rtc6705 ignores the CS
but as we are using it lets do it properly.